test: add collection page unit test
