+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import CollectionsPage from './pages/CollectionsPage';
+import CollectionsPage from './pages/CollectionPage';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import OrderPage from './pages/OrderPage';
